requirements.txt
requests==2.28.1
python-dotenv==0.20.0
redis==3.2.1
python-telegram-bot==13.*